Decouple Gradient.colourAt from colour range
Split Gradient.colourAt into two methods: colourIn, which gets the colour at a range index in a range of a specific length, and colourAt, which gets a colour at a specific palette index in a colour range.

Also, refactor getStopRangeIdxs and stopRangeIndex into methods that do not depend on a specific colour range.